refactor: auth 파일 구조 변경
- auth 폴더 삭제
